+++ examples/svs/low-level/only_core/main.go
@@ -76,9 +76,9 @@ loopCount:
 		// Receive code when available
 		case missing := <-recv:
 			for _, m := range missing {
-				for m.LowSeqno() <= m.HighSeqno() {
-					fmt.Println(m.Source().String() + ": " + strconv.FormatUint(m.LowSeqno(), 10))
-					m.Increment()
+				for m.LowSeq <= m.HighSeq {
+					fmt.Println(m.Dataset.String() + ": " + strconv.FormatUint(m.LowSeq, 10))
+					m.LowSeq++
 				}
 			}