+++ pkg/svs/onestate_core.go
@@ -177,7 +177,7 @@ func (c *oneStateCore) mergeVectorToLocal(vector StateVector) bool {
 	for p := vector.Entries().Back(); p != nil; p = p.Prev() {
 		lVal = c.local.Get(p.Kstr)
 		if lVal < p.Val {
-			missing = append(missing, NewMissingData(p.Kname, lVal+1, p.Val))
+			missing = append(missing, MissingData{Dataset: p.Kname, LowSeq: lVal + 1, HighSeq: p.Val})
 			c.local.Set(p.Kstr, p.Kname, p.Val, false)
 		} else if !slices.Contains(c.selfsets, p.Kstr) && lVal > p.Val {
 			isNewer = true