+++ docs/CHANGELOG.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ## Changed
+- De-interface small simple structures: `MissingData` and `StatusChange`.
 - Reorganized functions to match interface method order.
 - Changed naming of a constant variable to be more consistent with others that function the same.
 - Moved to go 1.22, updated all dependencies.
+
 
 ## [v0.0.0-alpha.14] - 2024-02-12
 ## Added